requirements.txt
openai~=1.66.3
PyPDF2~=3.0.1
fastapi~=0.115.11
python-dotenv~=1.0.1
python-multipart~=0.0.20
uvicorn~=0.34.0